update registrar to match example

+++ registrar/Registrar.cs
@@ -1,16 +1,13 @@
-﻿using Neo.SmartContract.Framework;
-using Neo.SmartContract.Framework.Services.Neo;
-using System.ComponentModel;
-using System.Reflection;    
+﻿using Neo.SmartContract.Framework.Services.Neo;
 
-[assembly: AssemblyDescription("Sample Smart Contract")]
-[assembly: AssemblyMetadata("ContractAuthor", "Harry Pierson")]
-[assembly: AssemblyMetadata("ContractEmail", "harrypierson@hotmail.com")]
-[assembly: AssemblyMetadata("ContractHasStorage", "true")]
+// [assembly: AssemblyDescription("Sample Smart Contract")]
+// [assembly: AssemblyMetadata("ContractAuthor", "Harry Pierson")]
+// [assembly: AssemblyMetadata("ContractEmail", "harrypierson@hotmail.com")]
+// [assembly: AssemblyMetadata("ContractHasStorage", "true")]
 
-namespace DevHawk.Neo.Experiments
+namespace Neo.SmartContract
 {
-    public class Registrar : SmartContract
+    public class Domain : Framework.SmartContract
     {
         public static object Main(string operation, params object[] args)
         {
@@ -29,39 +26,35 @@ namespace DevHawk.Neo.Experiments
             }
         }
 
-        [DisplayName("query")]
-        public static byte[] Query(string domain)
+        private static byte[] Query(string domain)
         {
             return Storage.Get(Storage.CurrentContext, domain);
         }
 
-        [DisplayName("register")]
-        public static bool Register(string domain, byte[] owner)
+        private static bool Register(string domain, byte[] owner)
         {
-            // if (!Runtime.CheckWitness(owner)) return false;
+            if (!Runtime.CheckWitness(owner)) return false;
             byte[] value = Storage.Get(Storage.CurrentContext, domain);
             if (value != null) return false;
             Storage.Put(Storage.CurrentContext, domain, owner);
             return true;
         }
 
-        [DisplayName("transfer")]
-        public static bool Transfer(string domain, byte[] to)
+        private static bool Transfer(string domain, byte[] to)
         {
-            // if (!Runtime.CheckWitness(to)) return false;
+            if (!Runtime.CheckWitness(to)) return false;
             byte[] from = Storage.Get(Storage.CurrentContext, domain);
             if (from == null) return false;
-            // if (!Runtime.CheckWitness(from)) return false;
+            if (!Runtime.CheckWitness(from)) return false;
             Storage.Put(Storage.CurrentContext, domain, to);
             return true;
         }
 
-        [DisplayName("delete")]
-        public static bool Delete(string domain)
+        private static bool Delete(string domain)
         {
             byte[] owner = Storage.Get(Storage.CurrentContext, domain);
             if (owner == null) return false;
-            // if (!Runtime.CheckWitness(owner)) return false;
+            if (!Runtime.CheckWitness(owner)) return false;
             Storage.Delete(Storage.CurrentContext, domain);
             return true;
         }
